Extract some common components from update subscription page
I am making a create subscription page and it will use these components
